Fixes issue with Step 1 Lasso
When creating the design matrix for glmnet, we previously used the same logical vector to filter out the treatment and response columns from all datasets. This ignored the fact that d0 and d1 (training data under A = 0 and A = 1) already had A removed via subset_trt whereas data still had A. With this change, we now remove the correct columns from data before using it for prediction.